Code changes for moving core.distances to lib.distances
