+++ package/MDAnalysis/analysis/contacts.py
@@ -149,7 +149,7 @@ import numpy
 import logging
 
 import MDAnalysis
-import MDAnalysis.core.distances
+import MDAnalysis.lib.distances
 from MDAnalysis.lib.util import openany
 
 
@@ -287,7 +287,7 @@ class ContactAnalysis(object):
 
         :Keywords:
            *results*
-              passed on to :func:`MDAnalysis.core.distances.self_distance_array`
+              passed on to :func:`MDAnalysis.lib.distances.self_distance_array`
               as a preallocated array
            *centroids*
               ``True``: calculate per-residue centroids from the selected atoms;
@@ -302,7 +302,7 @@ class ContactAnalysis(object):
         else:
             # centroids per residue (but only including the selected atoms)
             coordinates = numpy.array([residue.centroid() for residue in g.split("residue")])
-        return MDAnalysis.core.distances.self_distance_array(coordinates, **kwargs)
+        return MDAnalysis.lib.distances.self_distance_array(coordinates, **kwargs)
 
     def output_exists(self, force=False):
         """Return True if default output file already exists.
@@ -553,7 +553,7 @@ class ContactAnalysis1(object):
                                  (s, ref.atoms.numberOfAtoms(), sel.atoms.numberOfAtoms()))
 
         # compute reference contacts
-        dref = MDAnalysis.core.distances.distance_array(
+        dref = MDAnalysis.lib.distances.distance_array(
             self.references[0].coordinates(), self.references[1].coordinates())
         self.qref = self.qarray(dref)
         self.nref = self.qref.sum()
@@ -621,7 +621,7 @@ class ContactAnalysis1(object):
             for ts in self.universe.trajectory[start_frame:end_frame:step_value]:
                 frame = ts.frame
                 # use pre-allocated distance array to save a little bit of time
-                MDAnalysis.core.distances.distance_array(A.coordinates(), B.coordinates(), result=self.d)
+                MDAnalysis.lib.distances.distance_array(A.coordinates(), B.coordinates(), result=self.d)
                 self.qarray(self.d, out=self.q)
                 n1, q1 = self.qN(self.q, out=self._qtmp)
                 self.qavg += self.q
