+++ package/setup.py
@@ -231,7 +231,7 @@ if __name__ == '__main__':
                   define_macros=define_macros + parallel_macros,
                   extra_compile_args=parallel_args,
                   extra_link_args=parallel_args),
-        Extension("core.parallel.distances",
+        Extension("lib.parallel.distances",
                   ['src/numtools/distances_parallel.%s' % ("pyx" if use_cython else "c")],
                   include_dirs=include_dirs,
                   libraries=['m'] + parallel_libraries,
@@ -281,9 +281,9 @@ if __name__ == '__main__':
                     'MDAnalysis.coordinates.xdrfile',
                     'MDAnalysis.coordinates.pdb',
                     'MDAnalysis.core',
-                    'MDAnalysis.core.parallel',
                     'MDAnalysis.lib',
                     'MDAnalysis.lib.KDTree',
+                    'MDAnalysis.lib.parallel',
                     'MDAnalysis.selections',
                     'MDAnalysis.topology',
                     'MDAnalysis.topology.tpr',
